+++ src/app/layout.tsx
@@ -1,14 +1,14 @@
 import type { Metadata } from "next";
 import "./globals.css";
-import { ThemeProvider } from "@/components/theme-provider";
+import { ThemeProvider } from "@/components/provider/theme-provider";
 import React from "react";
 import Footer from "@/app/_components/footer";
 import NavBar from "@/app/_components/nav-bar";
-import ReactQueryProvider from "@/components/react-query-provider";
+import ReactQueryProvider from "@/components/provider/react-query-provider";
 import GoogleAd from "@/components/google-ad";
 import dynamic from "next/dynamic";
 import SearchBar from "@/app/_components/search-bar";
-import { DropdownProvider } from "@/components/dropdown-provider";
+import { DropdownProvider } from "@/components/provider/dropdown-provider";
 const LeaderBoardBox = dynamic(
   () => import("./_components/leader-board-box/index"),
   {
